requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
asgiref
attrs
contourpy
cycler
discord.py
dj-database-url
Django
django-email
django-environ
django-filter
djangorestframework
EasyProcess
entrypoint2
environ
fonttools
frozenlist
idna
kiwisolver
Markdown
matplotlib
MouseInfo
mss
multidict
networkx
numpy
packaging
Pillow
propcache
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyparsing
pyperclip
PyRect
pyscreenshot
PyScreeze
python-dateutil
pytweening
six
sqlparse
typing_extensions
tzdata
yarl
